+++ Runtime/Network/WebRequestVerb.cs
@@ -2,9 +2,12 @@ namespace Unity.Services.Authentication
 {
     enum WebRequestVerb
     {
+        Delete,
         Get,
+        Head,
+        Options,
+        Patch,
         Post,
-        Put,
-        Delete
+        Put
     }
 }