+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.5.0-pre.3] - 2023-03-01
+### Added
+- Added `GetPlayerNameAsync` to retrieve the current player name in the Authentication service
+- Added `UpdatePlayerNameAsync` to update the current player name in the Authentication service
+- Added `PlayerName` property and caching in the Authentication service
+### Changed
+- `PlayerId` property now properly returns null instead of an empty string by default
+- Package's license to refresh legal links.
+
 ## [2.4.0] - 2022-11-29
 ### Added
 - Added `SignInWithAppleGameCenterAsync`, `LinkWithAppleGameCenterAsync`, `UnlinkAppleGameCenterAsync`